+++ client/src/App.css
@@ -1,4 +1,5 @@
 @import "./static/styles/modal.css";
+@import "./static/styles/calendar.css";
 @import "./static/styles/search.css";
 @import "./static/styles/navbar.css";
 @import "./static/styles/footer.css";
